Manually place ships
The player can now manually place ships.
It shows where the new ship will be placed and allows the player to rotate the ship.
If the player attempts to place the ship over an existing ship, they will be told that is an invalid location and wont be able to place it there.