Bug fixes
Prevent player leaving the bounds of the grid while choosing where to fire.

When selecting a coordinate, a green line is displayed on either side of the selected one unless it was at the right hand side of the grid, where the right green line was not shown. This has been fixed.